requirements.txt
#
# This file is autogenerated by pip-compile with python 3.10
# To update, run:
#
#    pip-compile requirements.in
#
amqp==5.1.1
    # via kombu
appdirs==1.4.4
    # via black
asgiref==3.5.0
    # via django
attrs==21.4.0
    # via pytest
billiard==3.6.4.0
    # via celery
black==21.7b0
    # via -r requirements.in
celery==5.2.2
    # via
    #   -r requirements.in
    #   django-celery-results
cfgv==3.3.1
    # via pre-commit
click==8.1.2
    # via
    #   black
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
    #   pip-tools
click-didyoumean==0.3.0
    # via celery
click-plugins==1.1.1
    # via celery
click-repl==0.2.0
    # via celery
distlib==0.3.4
    # via virtualenv
django==4.0.4
    # via
    #   -r requirements.in
    #   django-cors-headers
    #   django-reset-migrations
    #   djangorestframework
django-celery-results==2.2.0
    # via -r requirements.in
django-cors-headers==3.8.0
    # via -r requirements.in
django-pytest==0.2.0
    # via -r requirements.in
django-reset-migrations==0.4.0
    # via -r requirements.in
djangorestframework==3.12.4
    # via -r requirements.in
filelock==3.6.0
    # via virtualenv
identify==2.4.12
    # via pre-commit
iniconfig==1.1.1
    # via pytest
isort==5.10.1
    # via -r requirements.in
kombu==5.2.4
    # via celery
mypy-extensions==0.4.3
    # via black
nodeenv==1.6.0
    # via pre-commit
packaging==21.3
    # via pytest
pathspec==0.9.0
    # via black
pep517==0.12.0
    # via pip-tools
pip-tools==6.6.0
    # via -r requirements.in
platformdirs==2.5.2
    # via virtualenv
pluggy==1.0.0
    # via pytest
pre-commit==2.18.1
    # via -r requirements.in
prompt-toolkit==3.0.29
    # via click-repl
py==1.11.0
    # via pytest
pyparsing==3.0.8
    # via packaging
pytest==7.1.1
    # via -r requirements.in
pytz==2022.1
    # via celery
pyyaml==6.0
    # via pre-commit
regex==2022.3.15
    # via black
six==1.16.0
    # via
    #   click-repl
    #   virtualenv
sqlparse==0.4.2
    # via django
toml==0.10.2
    # via pre-commit
tomli==1.2.3
    # via
    #   black
    #   pep517
    #   pytest
vine==5.0.0
    # via
    #   amqp
    #   celery
    #   kombu
virtualenv==20.14.1
    # via pre-commit
wcwidth==0.2.5
    # via prompt-toolkit
wheel==0.37.1
    # via pip-tools

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
